pyproject.toml
[tool.poetry]
name = "fastapi-course-app"
version = "0.1.0"
description = ""
authors = ["nshib <nshib54321@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.30.1"
fastapi = {extras = ["all"], version = "^0.111.0"}
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
bcrypt = "^4.1.3"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
celery = "^5.4.0"
pillow = "^10.4.0"
flower = "^2.0.1"
sqladmin = "^0.18.0"
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
pytest-cov = "^5.0.0"
black = "^24.8.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
pyright = "^1.1.380"
autoflake = "^2.3.1"
pre-commit = "^3.8.0"
pydantic-settings = "2.3.1"
python-json-logger = "^3.3.0"
sentry-sdk = {extras = ["fastapi"], version = "^2.35.0"}


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100

[tool.autoflake]
check_diff = true
imports = ["sqlalchemy", "app"]

[tool.pyright]
include = ["app"]
reportMissingImports = false
